extract `InstanceEnv::console_timer_end`
